requirements.txt
pandas>=2.0.0
dune-client>=1.2.0
pyyaml>=6.0
pandas-stubs>=2.3.3.251219